pyproject.toml
[tool.poetry]
name = "basty"
version = "0.1.0"
description = ""
authors = ["Ali Osman Berk Şapcı <aliosmanberk@sabanciuniv.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.0"
PyYAML = "6.0"
filterpy = "1.4.5"
hdbscan = "0.8.28"
joblib = "1.1.0"
pandas = "1.3.5"
scikit-learn = "1.0.2"
tqdm = "4.63.0"
umap-learn = "0.1.1"
numpy = "1.21.5"
scipy = "1.8.0"
PyWavelets = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
